feat: implement Landing Page with router
Implemented the Landing Page and integrated it with the router for navigation.
